Rename my_strings to constants
