+++ helpers.py
@@ -1,5 +1,5 @@
 import re
-import my_strings
+import constants
 
 def isInt(s):
     """
@@ -42,7 +42,7 @@ def reply_and_upvote(c, response, respond=False):
     """
     if respond:
         print(f'Responding to comment{c.permalink}')
-        c.reply(response + my_strings.TAIL)
+        c.reply(response + constants.TAIL)
         c.upvote()
     else:
         print(f'Not responding to comment {c.permalink}')
